Migrated multiplayer to Ship class.
The Ship class now includes the vast majority of the functionality it
needs to work as intended. Now, only player bounds and nrg visual update
are handled in the SpaceFights update method. Projectile creation,
update and drawing are now all handled by their source ship. 

TODO FOR PROTOTYPE: We still need to create a wall or two and build
collision around those for the players.