Made ShipType Enum for easier ship class creation
When calling the Ship constructor, now only need to supply two
parameters: ShipType and an Integer representing the player number.
ShipType is a custom type that is defined by the Enum ShipType, and
currently includes Rhino and Vulture. More can be added, as can more
fields to further differentiate the various classes of ship

Change-Id: Ib1e0a87596c6c09302cbb273205412d539c4cdc7